pyproject.toml
[project]
name = "message-triange"
version = "1"
description = ""
authors = [
    {name = "Badmus Owolabi"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "textblob (>=0.19.0,<0.20.0)",
    "celery (>=5.5.3,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
pytest = "^8.4.1"
pytest-django = "^4.11.1"

